requirements.txt
scrapy
selenium
mysql-connector-python
mysqlx-connector-python
scrapy-splash
fastapi
uvicorn
SQLAlchemy
pymysql
cryptography
pyjwt
